Create documentation and releaseNotes subdirectories and rename Explore components
