+++ src/Routes.tsx
@@ -13,7 +13,7 @@ import {
   BetaPage,
   BetaSuccess,
   DisabledApplyForm,
-  ExploreDocs,
+  DocumentationPage,
   Home,
   News,
   OAuth,
@@ -55,7 +55,7 @@ export function topLevelRoutes(props: RouteComponentProps<void>) {
         <Route path="/applied" component={ApplySuccess} />
         <Route path="/beta" component={BetaPage} />
         <Route path="/beta-success" component={BetaSuccess} />
-        <Route path="/explore/:apiCategoryKey?" component={ExploreDocs} />
+        <Route path="/explore/:apiCategoryKey?" component={DocumentationPage} />
         <Route exact={true} path="/explore/:apiCategoryKey/docs/:apiName" />
         <Route path="/oauth" component={OAuth} />
         <Route path="/release-notes/:apiCategoryKey?" component={ReleaseNotes} />
