+++ src/containers/documentation/ApiKeyAuth.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Redirect } from 'react-router-dom';
 
-import { IApiNameParam } from '../types';
+import { IApiNameParam } from '../../types';
 
 export class ApiKeyAuth extends React.Component<IApiNameParam, {}> {
   public render() {
