+++ src/containers/documentation/QuickstartPage.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 
 import { Redirect, RouteComponentProps } from 'react-router';
 
-import { apiDefs } from '../apiDefs';
-import QuickstartWrapper from '../components/QuickstartWrapper';
-import { IApiNameParam } from '../types';
+import { apiDefs } from '../../apiDefs';
+import QuickstartWrapper from '../../components/QuickstartWrapper';
+import { IApiNameParam } from '../../types';
 
 export default class QuickstartPage extends React.Component<RouteComponentProps<IApiNameParam>, {}> {
   public render() {
