+++ src/containers/documentation/swaggerPlugins/VersionSelect.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 #version-select {
   select {
