+++ src/containers/releaseNotes/CategoryReleaseNotesPage.tsx
@@ -3,11 +3,11 @@ import * as React from 'react';
 import { Flag } from 'flag';
 import { RouteComponentProps } from 'react-router';
 
-import { apiDefs, IApiDescription } from '../apiDefs';
-import CardLink from '../components/CardLink';
-import PageHeader from '../components/PageHeader';
-import { VAInternalOnlyTag } from '../components/VAInternalOnlyTag';
-import { IApiNameParam } from '../types';
+import { apiDefs, IApiDescription } from '../../apiDefs';
+import CardLink from '../../components/CardLink';
+import PageHeader from '../../components/PageHeader';
+import { VAInternalOnlyTag } from '../../components/VAInternalOnlyTag';
+import { IApiNameParam } from '../../types';
 
 export default class CategoryReleaseNotesPage extends React.Component<RouteComponentProps<IApiNameParam>, {}> {
   public render() {
