+++ src/containers/releaseNotes/ReleaseNotesOverview.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import { apiCategoryOrder, apiDefs } from '../apiDefs';
-import CardLink from '../components/CardLink';
-import PageHeader from '../components/PageHeader';
+import { apiCategoryOrder, apiDefs } from '../../apiDefs';
+import CardLink from '../../components/CardLink';
+import PageHeader from '../../components/PageHeader';
 
 function cards(parent: string): any {
   return apiCategoryOrder.map((apiCategoryKey: string) => {
